Refactor: Separate notification into component
